+++ test/lint.js
@@ -1,5 +1,5 @@
 const assert = require("assert");
-const { lintStr } = require("../lint.js");
+const { lintStr } = require("../src/lint.js");
 
 describe("Linter", () => {
   describe("lintStr()", () => {
@@ -145,91 +145,6 @@ describe("Linter", () => {
       assert.equal(
         lintStr(toStr(specs)),
         toStr(["https://www.w3.org/TR/duplicate/"]));
-    });
-
-    it("throws if specs is not an array", () => {
-      const specs = 0;
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^specs should be array$/);
-    });
-
-    it("throws if specs is an empty array", () => {
-      const specs = [];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^specs should NOT have fewer than 1 items$/);
-    });
-
-    it("throws if specs contains an item with a wrong type", () => {
-      const specs = [0];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^specs\[0\] should be a string or an object$/);
-    });
-
-    it("throws if specs contains an invalid URL", () => {
-      const specs = ["https://?"];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^TypeError (.*) Invalid URL/);
-    });
-
-    it("throws if specs contains an object without URL", () => {
-      const specs = [{}];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^specs\[0\] should have required property 'url'$/);
-    });
-
-    it("throws if specs contains an object with an invalid URL", () => {
-      const specs = [{ url: "invalid" }];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^specs\[0\]\.url should match format "uri"$/);
-    });
-
-    it("throws if specs contains an object with additional properties", () => {
-      const specs = [{ url: "https://example.org/", invalid: "test" }];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^specs\[0\] should not have additional property 'invalid'$/);
-    });
-
-    it("throws when no name can be derived from a URL", () => {
-      const specs = ["https://example.org/"];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^Cannot extract meaningful name from/);
-    });
-
-    it("throws when a delta spec does not have a full previous level", () => {
-      const specs = ["https://www.w3.org/TR/spec/ delta"];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^Delta spec\(s\) found without full previous level/);
-    });
-
-    it("throws when a delta spec is flagged as current", () => {
-      const specs = [
-        { url: "https://www.w3.org/TR/spec-1/" },
-        { url: "https://www.w3.org/TR/spec-2/",
-          seriesComposition: "delta", forceCurrent: true }
-      ];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^Delta spec\(s\) found that are also flagged as current/);
-    });
-
-    it("throws when multiple levels are defined as current", () => {
-      const specs = [
-        "https://www.w3.org/TR/spec-1/ current",
-        "https://www.w3.org/TR/spec-2/ current",
-        "https://www.w3.org/TR/spec-3/",
-      ];
-      assert.throws(
-        () => lintStr(toStr(specs)),
-        /^Too many current specs for shortname\(s\)/);
     });
   });
 });
